+++ CHANGELOG.md
@@ -9,4 +9,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Added
 
-- sample demo app for fj-daogen version 1.3.3
+- sample demo app for fj-daogen version 1.5.0